+++ src/components/index.ts
@@ -1,7 +1,6 @@
 import NotFound from '../pages/notFound/NotFound.tsx';
 import Header from './header/Header.tsx';
 import Layout from './layout/Layout.tsx';
-import LoginRegisterLayout from './layout/login-register-layout.tsx';
 import Loading from './loading/Loading.tsx';
 import AuthInput from './login_register_inputs/auth_input.tsx';
 import FormInput from './login_register_inputs/form_input.tsx';
@@ -16,7 +15,6 @@ export {
   Header,
   Layout,
   Loading,
-  LoginRegisterLayout,
   Logo,
   Navigation,
   NotFound,
