+++ src/components/login_register_inputs/form_input.tsx
@@ -15,13 +15,13 @@ import dayjs from 'dayjs';
 import { lazy, useState } from 'react';
 import { Controller, useFormContext, useWatch } from 'react-hook-form';
 
+import { RegisterInputProps } from '../../pages/login-register/register/interfaces';
+
 const DatePicker = lazy(() =>
   import('@mui/x-date-pickers').then(module => ({
     default: module.DatePicker,
   })),
 );
-
-import { RegisterInputProps } from '../../pages/register/interfaces';
 
 interface FormInputProps extends Partial<RegisterInputProps> {
   name: string;
