+++ src/pages/login-register/login/Login.tsx
@@ -16,17 +16,16 @@ import { SubmitHandler, useForm } from 'react-hook-form';
 import { useNavigate } from 'react-router';
 import { toast } from 'react-toastify';
 
-import { Loading } from '../../components';
-import { useAppDispatch, useAppSelector } from '../../redux/hooks';
-import { getCustomer } from '../../redux/selectors';
-import loginCustomer from '../../redux/slices/asyncThunks/loginCustomer';
+import { Loading } from '../../../components';
+import { useAppDispatch, useAppSelector } from '../../../redux/hooks';
+import { getCustomer } from '../../../redux/selectors';
+import loginCustomer from '../../../redux/slices/asyncThunks/loginCustomer';
 import {
   useGetActiveCartQuery,
   useUpdateCartMutation,
-  // useMergeCartMutation,
-} from '../../services/api';
-import { ECartUpdateActions } from '../../services/interfaces/updateCart.interface';
-import CONSTANTS from '../../utils/CONSTANTS';
+} from '../../../services/api';
+import { ECartUpdateActions } from '../../../services/interfaces/updateCart.interface';
+import CONSTANTS from '../../../utils/CONSTANTS';
 import schema from './login_schema';
 
 interface IFormInputs {
