+++ src/pages/profile/auth-layout/password-changer/passwordSchema.ts
@@ -1,6 +1,6 @@
 import * as yup from 'yup';
 
-import { noSpacesRegex } from '../../../register/interfaces';
+import { noSpacesRegex } from '../../../login-register/register/interfaces';
 
 const passwordSchema = yup.object().shape({
   oldPassword: yup
