+++ src/pages/profile/auth-layout/profile-form/profileFormSchema.ts
@@ -1,6 +1,9 @@
 import * as yup from 'yup';
 
-import { domainRegex, onlyLettersRegex } from '../../../register/interfaces';
+import {
+  domainRegex,
+  onlyLettersRegex,
+} from '../../../login-register/register/interfaces';
 
 export const profileFormSchema = yup.object().shape({
   firstName: yup
